requirements_windows.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
amqp==5.1.1
aniso8601==9.0.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.2
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
ffmpeg==1.4
filelock==3.12.2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-BabelEx==0.9.4
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
fonttools==4.42.1
frozenlist==1.4.0
greenlet==2.0.2
gTTS==2.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwt==1.3.1
kiwisolver==1.4.4
kombu==5.3.1
lazy_loader==0.3
librosa==0.10.1
lit==16.0.6
llvmlite==0.40.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.8
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.2
sentry-sdk==1.29.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
speaklater==1.3
SQLAlchemy==2.0.20
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
tiktoken==0.3.3
torch==1.12.1
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
